feat: Add new cpus_per_core metric
