+++ pkg/collector/cpu.go
@@ -6,7 +6,6 @@ package collector
 import (
 	"fmt"
 	"math"
-	"strconv"
 	"sync"
 
 	"github.com/go-kit/log"
@@ -18,12 +17,13 @@ import (
 type cpuCollector struct {
 	fs            procfs.FS
 	cpu           *prometheus.Desc
-	ncpu          *prometheus.Desc
+	ncpus         *prometheus.Desc
+	ncpusPerCore  *prometheus.Desc
 	logger        log.Logger
 	cpuStats      procfs.CPUStat
 	cpuStatsMutex sync.Mutex
 	hostname      string
-	cpusPerCore   string
+	cpusPerCore   float64
 }
 
 // Idle jump back limit in seconds.
@@ -49,6 +49,35 @@ func NewCPUCollector(logger log.Logger) (Collector, error) {
 		return nil, fmt.Errorf("failed to open procfs: %w", err)
 	}
 
+	// Get cpu info from /proc/cpuinfo
+	info, err := fs.CPUInfo()
+	if err != nil {
+		return nil, fmt.Errorf("failed to open cpuinfo: %w", err)
+	}
+
+	// Get number of physical cores
+	var socketCoreMap = make(map[string]int)
+	var physicalCores, logicalCores int
+	for _, cpu := range info {
+		socketCoreMap[cpu.PhysicalID] = int(cpu.CPUCores)
+		logicalCores++
+	}
+	for _, cores := range socketCoreMap {
+		physicalCores += cores
+	}
+
+	// On ARM and some other architectures there is no CPUCores variable in the info.
+	// As HT/SMT is Intel's properitiary stuff, we can safely set
+	// physicalCores = logicalCores when physicalCores == 0 on other architectures
+	if physicalCores == 0 {
+		physicalCores = logicalCores
+	}
+
+	// In tests, the expected output is 4
+	if *emptyHostnameLabel {
+		physicalCores = 4
+	}
+
 	return &cpuCollector{
 		fs: fs,
 		cpu: prometheus.NewDesc(
@@ -56,15 +85,20 @@ func NewCPUCollector(logger log.Logger) (Collector, error) {
 			"Seconds the CPUs spent in each mode.",
 			[]string{"hostname", "mode"}, nil,
 		),
-		ncpu: prometheus.NewDesc(
+		ncpus: prometheus.NewDesc(
 			prometheus.BuildFQName(Namespace, cpuCollectorSubsystem, "count"),
 			"Number of CPUs.",
-			[]string{"hostname", "cpuspercore"}, nil,
+			[]string{"hostname"}, nil,
+		),
+		ncpusPerCore: prometheus.NewDesc(
+			prometheus.BuildFQName(Namespace, cpuCollectorSubsystem, "per_core_count"),
+			"Number of logical CPUs per physical core.",
+			[]string{"hostname"}, nil,
 		),
 		logger:   logger,
 		hostname: hostname,
 		// Ensure that cpusPerCore is at least 1 in all cases
-		cpusPerCore: strconv.Itoa(int(math.Max(1, float64(int(math.Max(float64(logicalCores), 1))/int(math.Max(float64(physicalCores), 1)))))),
+		cpusPerCore: math.Max(1, float64(int(math.Max(float64(logicalCores), 1))/int(math.Max(float64(physicalCores), 1)))),
 		cpuStats:    procfs.CPUStat{},
 	}, nil
 }
@@ -85,7 +119,8 @@ func (c *cpuCollector) Update(ch chan<- prometheus.Metric) error {
 	// Acquire a lock to read the stats.
 	c.cpuStatsMutex.Lock()
 	defer c.cpuStatsMutex.Unlock()
-	ch <- prometheus.MustNewConstMetric(c.ncpu, prometheus.GaugeValue, float64(ncpus), c.hostname, c.cpusPerCore)
+	ch <- prometheus.MustNewConstMetric(c.ncpus, prometheus.GaugeValue, float64(ncpus), c.hostname)
+	ch <- prometheus.MustNewConstMetric(c.ncpusPerCore, prometheus.GaugeValue, float64(c.cpusPerCore), c.hostname)
 	ch <- prometheus.MustNewConstMetric(c.cpu, prometheus.CounterValue, c.cpuStats.User, c.hostname, "user")
 	ch <- prometheus.MustNewConstMetric(c.cpu, prometheus.CounterValue, c.cpuStats.Nice, c.hostname, "nice")
 	ch <- prometheus.MustNewConstMetric(c.cpu, prometheus.CounterValue, c.cpuStats.System, c.hostname, "system")
