+++ scripts/checkmetrics.sh
@@ -9,7 +9,7 @@ fi
 search_dir="$2"
 for entry in "$search_dir"/*
 do
-  lint=$($1 check metrics < "$entry" 2>&1 | grep -v -E "^ceems_compute_unit_(memory_fail_count|memsw_fail_count)|ceems_meminfo_|ceems_cpu_count")
+  lint=$($1 check metrics < "$entry" 2>&1 | grep -v -E "^ceems_compute_unit_(memory_fail_count|memsw_fail_count)|ceems_meminfo_|ceems_cpu_count|ceems_cpu_per_core_count")
 
   if [[ -n $lint ]]; then
       echo -e "Some Prometheus metrics do not follow best practices:\n"
